requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
bootstrap4==0.1.0
celery==5.0.2
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.3
django-crispy-forms==1.9.2
djangorestframework==3.12.2
gunicorn==20.0.4
honcho==1.0.1
kombu==5.0.2
lml==0.1.0
prompt-toolkit==3.0.8
psycopg2==2.8.6
pyexcel-io==0.6.4
pyexcel-xls==0.6.1
pytz==2020.4
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
